requirements.txt
# Core Dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.3.0
plotly>=5.15.0
PyYAML>=6.0
seaborn>=0.12.0
matplotlib>=3.7.0
altair>=5.0.0
streamlit-aggrid>=0.3.0
streamlit-option-menu>=0.3.0
Pillow>=10.0.0
openpyxl>=3.1.0
faker>=20.0.0

# Advanced ML and AI
tensorflow>=2.13.0
torch>=2.0.0
xgboost>=1.7.0
lightgbm>=4.0.0
catboost>=1.2.0
mlflow>=2.5.0
optuna>=3.2.0

# Deep Learning and Neural Networks
keras>=2.13.0
requests>=2.31.0
httpx>=0.24.0
python-dotenv>=1.0.0
joblib>=1.3.0

# API & web
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-multipart>=0.0.6
jinja2>=3.1.0

# Auth (future-ready)
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0

# Data I/O & viz extras
openpyxl>=3.1.0
xlsxwriter>=3.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Background & cache (optional)
celery>=5.3.0
redis>=4.5.0

# Monitoring & logging
prometheus-client>=0.17.0
structlog>=23.0.0

# Database & ORM (optional future Postgres path)
SQLAlchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
pymongo>=4.5.0

# Testing & quality
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Performance & load (optional)
locust>=2.16.0
fakeredis>=2.18.0
notebook>=7.0.0
debugpy>=1.6.0

# Security and Authentication
cryptography>=41.0.0
python-jose>=3.3.0
python-multipart>=0.0.6

# Utilities and Helpers
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.4.0
tqdm>=4.65.0
pydantic>=2.0.0
marshmallow>=3.20.0
joblib>=1.3.0

# Data Processing and Analysis
dask>=2023.6.0
vaex>=4.16.0
modin>=0.20.0
polars>=0.19.0

# Visualization and Dashboard
bokeh>=3.2.0
dash>=2.11.0
gradio>=3.35.0
panel>=1.2.0

# Time Series and Forecasting
prophet>=1.1.4
statsmodels>=0.14.0
arch>=6.2.0
pyflux>=0.4.18

# Natural Language Processing
spacy>=3.6.0
nltk>=3.8.1
textblob>=0.17.1
gensim>=4.3.0

# Geospatial and Location
geopandas>=0.13.0
folium>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0 


"""
# set "MODEL_PATH=models\model.pkl"
# uvicorn src.ml.model_server:app --host 0.0.0.0 --port 8001

# uvicorn src.api.gateway:app --host 0.0.0.0 --port 8002

# set "GATEWAY_URL=http://localhost:8002"
# streamlit run .\main.py
"""